renamed IT package to integration under tests
