+++ src/test/resources/com/google/jenkins/plugins/storage/integration/downloadstep_test.txt
