+++ src/test/resources/com/google/jenkins/plugins/storage/integration/malformedDownloadStepPipeline.groovy
